+++ docs/whats_new.rst
@@ -36,7 +36,8 @@ Enhancements
 - Add type hints to datasets (:gh:`590` by `Pierre Guetschel`_)
 - Add channel names and montage to  :class:`braindecode.datasets.TUH` (:gh:`593` by `Pierre Guetschel`_)
 - Add ``offset`` arg to :function:`braindecode.preprocessing.preprocess` (:gh:`599` by `Pierre Guetschel`_)
-
+- Add type hints to preprocessing (:gh:`600` by `Pierre Guetschel`_)
+- Add ``mypy`` type checks to pre-commit and CI (:gh:`606` by `Pierre Guetschel`_)
 
 Bugs
 ~~~~
@@ -44,11 +45,15 @@ Bugs
 - Fix deprecated test and changing the what's new checker (:gh: 569 by `Bruno Aristimunha`_)
 - Fix issue with coverage CI and adding a condition on the test for avoid HTML errors (:gh: 591 by `Bruno Aristimunha`_)
 - Constraint the version of mne (:gh: 594 by `Bruno Aristimunha`_)
+- Fix type errors (:gh:`606` by `Pierre Guetschel`_)
+- Warn when applying preprocessing steps on a :class:`braindecode.datasets.base.EEGWindowsDataset` (:gh:`607` by `Pierre Guetschel`_)
 - Fix dead link & Sphinx URL formatting (:gh:`602` by `John Muradeli`_)
 - Fix `tmax` in `SleepPhysionet` (:gh:`612` by `John Muradeli`_)
 
 API changes
 ~~~~~~~~~~~
+- Expose the ``use_mne_epochs parameter`` of :function:`braindecode.preprocessing.create_windows_from_events` (:gh:`607` by `Pierre Guetschel`_)
+
 
 .. _changes_0_8_0:
 Current 0.8 (11-2022)